Update to redacted source state
This commit is a NON-FUNCTIONAL display commit of
the source that was used in the course submission. It
is redacted and waiting for determining the Xilinx
licensing terms.

A functional Docker image is up for works, if the relevant
parts can be displayed and the time and will is there.